add checklist service + code refactoring in task service and separation of business logic

+++ src/main/java/ru/sushchenko/trelloclone/service/impl/ChecklistServiceImpl.java
@@ -8,9 +8,12 @@ import ru.sushchenko.trelloclone.dto.checklist.ChecklistRequest;
 import ru.sushchenko.trelloclone.dto.checklist.ChecklistResponse;
 import ru.sushchenko.trelloclone.entity.Checklist;
 import ru.sushchenko.trelloclone.entity.Task;
+import ru.sushchenko.trelloclone.entity.User;
 import ru.sushchenko.trelloclone.repo.ChecklistRepo;
 import ru.sushchenko.trelloclone.service.ChecklistService;
-import ru.sushchenko.trelloclone.utils.exception.ChecklistNotFoundException;
+import ru.sushchenko.trelloclone.service.TaskService;
+import ru.sushchenko.trelloclone.utils.exception.NotEnoughPermissionsException;
+import ru.sushchenko.trelloclone.utils.exception.ResourceNotFoundException;
 import ru.sushchenko.trelloclone.utils.mapper.ChecklistMapper;
 
 import java.util.HashSet;
@@ -24,14 +27,22 @@ import java.util.stream.Collectors;
 public class ChecklistServiceImpl implements ChecklistService {
     private final ChecklistRepo checklistRepo;
     private final ChecklistMapper checklistMapper;
+    private final TaskService taskService;
 
     @Override
     @Transactional
-    public ChecklistResponse addChecklist(ChecklistRequest checklistDto, Task task) {
-        Checklist checklist = checklistMapper.toEntity(checklistDto);
-        checklist.setCheckItems(new HashSet<>());
-        checklist.setTask(task);
-        return checklistMapper.toDto(checklistRepo.save(checklist));
+    public ChecklistResponse addChecklistToTask(UUID taskId, ChecklistRequest checklistDto, User currentUser) {
+        Task task = taskService.getExistingTask(taskId);
+        if(taskService.checkIfAllowedToModifyTask(task, currentUser)) {
+            Checklist checklist = checklistMapper.toEntity(checklistDto);
+            checklist.setCheckItems(new HashSet<>());
+            checklist.setTask(task);
+            Checklist savedChecklist = checklistRepo.save(checklist);
+            log.info("Checklist with id: {} added for task with id: {}", savedChecklist.getId(), taskId);
+            return checklistMapper.toDto(savedChecklist);
+        } else {
+            throw new NotEnoughPermissionsException(currentUser.getId(), taskId);
+        }
     }
 
     @Override
@@ -42,21 +53,32 @@ public class ChecklistServiceImpl implements ChecklistService {
     }
 
     @Override
-    @Transactional
-    public void deleteChecklistById(UUID id) {
-        checklistRepo.deleteById(id);
+    public void deleteChecklistById(UUID id, User currentUser) {
+        Checklist checklist = getExistingChecklist(id);
+        Task task = checklist.getTask();
+        if(taskService.checkIfAllowedToModifyTask(task, currentUser)) {
+            checklistRepo.deleteById(id);
+            log.info("Checklist with id: {} deleted by user with id: {}", id, currentUser.getId());
+        } else {
+            throw new NotEnoughPermissionsException(currentUser.getId(), id);
+        }
     }
 
     @Override
-    @Transactional
-    public ChecklistResponse updateChecklistById(UUID checklistId, ChecklistRequest checklistDto) {
-        Checklist checklist = getExistingChecklist(checklistId);
-        checklistMapper.mergeEntityToDto(checklistDto, checklist);
-        Checklist updatedChecklist = checklistRepo.save(checklist);
-        return checklistMapper.toDto(updatedChecklist);
+    public ChecklistResponse updateChecklistById(UUID id, ChecklistRequest checklistDto, User currentUser) {
+        Checklist checklist = getExistingChecklist(id);
+        Task task = checklist.getTask();
+        if(taskService.checkIfAllowedToModifyTask(task, currentUser)) {
+            checklistMapper.mergeEntityToDto(checklistDto, checklist);
+            Checklist updatedChecklist = checklistRepo.save(checklist);
+            log.info("Checklist with id: {} updated by user with id: {}", id, currentUser.getId());
+            return checklistMapper.toDto(updatedChecklist);
+        } else {
+            throw new NotEnoughPermissionsException(currentUser.getId(), id);
+        }
     }
 
     private Checklist getExistingChecklist(UUID id) {
-        return checklistRepo.findById(id).orElseThrow(() -> new ChecklistNotFoundException(id));
+        return checklistRepo.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
     }
 }
